correct ngrok command
with the current command I got the error:
`ERROR:  unknown shorthand flag: 'o' in -ost-header=yves.de.spryker.local:443ngrok`
even if the "h" is added, this was always the output.
While with the command:
`ngrok http yves.de.spryker.local:443` the dashboard was displayed.  
This command will forward to our local Spryker machine to test. Is not necessary to rewrite the host header if it was not added locally before.